fixed quantity parsing

+++ flutter_app/lib/models/recipe_ingredient_amount.dart
@@ -8,7 +8,7 @@ class RecipeIngredientAmount {
       RecipeIngredientAmount();
 
   @JsonKey(name: 'is_approximate') bool isApproximate;
-  int quantity;
+  double quantity;
   @JsonKey(name: 'quantity_up_to') dynamic quantityUpTo;
   String text;
   String unit;
